# lib/api-zod/src/generated/types/adminAssessmentRow.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindCheck AI - Mental Health Risk & Mood Pattern Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAssessmentRow {
  id: number;
  user_id: number;
  user_name: string;
  user_email: string;
  risk_level: string;
  digital_wellbeing_score: number;
  mood_score: number;
  created_at: Date;
}