# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindCheck AI - Mental Health Risk & Mood Pattern Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputExerciseFrequency } from './assessmentInputExerciseFrequency';
import type { AssessmentInputMood } from './assessmentInputMood';
import type { AssessmentInputSleepQuality } from './assessmentInputSleepQuality';
import type { AssessmentInputSocialComparison } from './assessmentInputSocialComparison';
import type { AssessmentInputSocialMediaHours } from './assessmentInputSocialMediaHours';
import type { AssessmentInputStressFrequency } from './assessmentInputStressFrequency';

export interface AssessmentInput {
  social_media_hours: AssessmentInputSocialMediaHours;
  sleep_quality: AssessmentInputSleepQuality;
  stress_frequency: AssessmentInputStressFrequency;
  social_comparison: AssessmentInputSocialComparison;
  exercise_frequency: AssessmentInputExerciseFrequency;
  mood: AssessmentInputMood;
}