# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindCheck AI - Mental Health Risk & Mood Pattern Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from './trendPoint';

export interface DashboardSummary {
  total_assessments: number;
  /** @nullable */
  latest_risk_level: string | null;
  /** @nullable */
  latest_digital_wellbeing_score: number | null;
  /** @nullable */
  latest_mood_score: number | null;
  risk_trend: TrendPoint[];
  mood_trend: TrendPoint[];
  wellbeing_trend?: TrendPoint[];
}