# lib/api-zod/src/generated/types/featureImportance.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindCheck AI - Mental Health Risk & Mood Pattern Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureImportance {
  feature: string;
  importance: number;
  label: string;
}